Move vb code around
