+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/LanguageServices/RemoveUnnecessaryImports/AbstractRemoveUnnecessaryImportsService.cs
@@ -25,22 +25,6 @@ namespace Microsoft.CodeAnalysis.RemoveUnnecessaryImports
 
         public abstract Task<Document> RemoveUnnecessaryImportsAsync(Document fromDocument, Func<SyntaxNode, bool>? predicate, SyntaxFormattingOptions? formattingOptions, CancellationToken cancellationToken);
 
-        protected static SyntaxToken StripNewLines(ISyntaxFacts syntaxFacts, SyntaxToken token)
-        {
-            var trimmedLeadingTrivia = token.LeadingTrivia.SkipWhile(syntaxFacts.IsEndOfLineTrivia).ToList();
-
-            // If the list ends with 3 newlines remove the last one until there's only 2 newlines to end the leading trivia.
-            while (trimmedLeadingTrivia.Count >= 3 &&
-                   syntaxFacts.IsEndOfLineTrivia(trimmedLeadingTrivia[^3]) &&
-                   syntaxFacts.IsEndOfLineTrivia(trimmedLeadingTrivia[^2]) &&
-                   syntaxFacts.IsEndOfLineTrivia(trimmedLeadingTrivia[^1]))
-            {
-                trimmedLeadingTrivia.RemoveAt(trimmedLeadingTrivia.Count - 1);
-            }
-
-            return token.WithLeadingTrivia(trimmedLeadingTrivia);
-        }
-
         protected async Task<HashSet<T>> GetCommonUnnecessaryImportsOfAllContextAsync(
             Document document, Func<SyntaxNode, bool> predicate, CancellationToken cancellationToken)
         {
